chore: rename mock mode button file and move to mocks folder
