+++ packages/trip-form/src/MetroModeSelector/AdvancedModeSettingsButton.story.tsx
@@ -12,7 +12,7 @@ import {
   defaultModeButtonDefinitions,
   getIcon,
   modeSettingDefinitionsWithDropdown
-} from "./mockButtons";
+} from "../__mocks__/mode-selector-buttons";
 
 const initialState = {
   enabledModeButtons: ["transit"],
