+++ packages/trip-form/src/__mocks__/mode-selector-buttons.tsx
